* master: (79 commits)
  Stop using internal array/object pointer in foreach by reference according to https://wiki.php.net/rfc/php7_foreach
  Fix "forech" statemt behaviour according to https://wiki.php.net/rfc/php7_foreach
  Pass maxlifetime to save handlers
  Added test for #69017
  Fixed bug #69017 (Fail to push to the empty array with the constant value defined in class scope)
  Align format to Fixed bug #... , later used to create the website changelog
  Fixed bug #69025 (Invalid read of size 4 when calling __callStatic)
  add missing test for class member access with deference relates to https://wiki.php.net/rfc/uniform_variable_syntax
  fix dir separator in test
  Add missing NEWS entry for revert of bug #41631
  Add CVE ID to bug #68676
  Align email format for recent NEWS entries
  Bug #55508 was fixed in 5.5.19 with 15ba757a, not in 5.5.18
  Add missing NEWS entry for revert of bug #41631
  Add missing NEWS entry for bug #68027
  Align format for bug #68799
  Bug #68361 was fixed in 5.5.20 using 327d4f9a, not in 5.5.19
  Bug #55618 was fixed in 5.5.21 using eaf107c0
  fix incompatible pointer type
  fix values for gid and uid
  ...

Conflicts:
	Zend/zend_hash.c
	Zend/zend_types.h
	Zend/zend_vm_def.h
	Zend/zend_vm_execute.h
	ext/standard/array.c

+++ ext/json/json.c
@@ -191,7 +191,7 @@ PHP_JSON_API void php_json_decode_ex(zval *return_value, char *str, size_t str_l
 {
 	php_json_parser parser;
 
-	php_json_parser_init(&parser, return_value, str, str_len, options, depth);
+	php_json_parser_init(&parser, return_value, str, str_len, (int)options, (int)depth);
 
 	if (php_json_yyparse(&parser)) {
 		JSON_G(error_code) = php_json_parser_error_code(&parser);
@@ -215,9 +215,9 @@ static PHP_FUNCTION(json_encode)
 
 	JSON_G(error_code) = PHP_JSON_ERROR_NONE;
 
-	JSON_G(encode_max_depth) = depth;
+	JSON_G(encode_max_depth) = (int)depth;
 
-	php_json_encode(&buf, parameter, options);
+	php_json_encode(&buf, parameter, (int)options);
 
 	if (JSON_G(error_code) != PHP_JSON_ERROR_NONE && !(options & PHP_JSON_PARTIAL_OUTPUT_ON_ERROR)) {
 		smart_str_free(&buf);
